refactor: split demo chart spec
