+++ packages/vchart-theme-demo-component/src/spec/funnel.ts
@@ -1,0 +1,39 @@
+export const funnelSpec = {
+  type: 'funnel',
+  categoryField: 'name',
+  valueField: 'value',
+  data: [
+    {
+      id: 'funnel',
+      values: [
+        {
+          value: 100,
+          name: 'Step1'
+        },
+        {
+          value: 80,
+          name: 'Step2'
+        },
+        {
+          value: 60,
+          name: 'Step3'
+        },
+        {
+          value: 40,
+          name: 'Step4'
+        },
+        {
+          value: 20,
+          name: 'Step5'
+        }
+      ]
+    }
+  ],
+  label: {
+    visible: true
+  },
+  legends: {
+    visible: true,
+    orient: 'bottom'
+  }
+};
